+++ sorted_nearest/src/windows.pyx
@@ -84,7 +84,7 @@ cpdef makewindows64(const long [::1] indexes, const long [::1] starts, const lon
         int nfound = 0
         int i = 0
         int length = len(starts)
-        int start, end
+        long start, end
 
     output_arr_indexes = np.ones(max_n_windows, dtype=long) * -1
     output_arr_start = np.ones(max_n_windows, dtype=long) * -1
